interface: fixed conditional to only show avatar in chat when we have shared it
