+++ pkg/interface/src/views/apps/chat/components/ShareProfile.js
@@ -7,69 +7,15 @@ import { Box, Row, Text, BaseImage } from '@tlon/indigo-react';
 import { uxToHex } from '~/logic/lib/util';
 import { Sigil } from '~/logic/lib/sigil';
 
-const pathAsResource = (path) => {
-  if (!path) {
-    return false;
-  }
-  const pathArr = path.split('/');
-  if (pathArr.length !== 4) {
-    return false;
-  }
-
-  return {
-    entity: pathArr[2],
-    name: pathArr[3]
-  };
-};
-
 export const ShareProfile = (props) => {
-  const { api, hideBanner, group, groupPath } = props;
-
-  const [showBanner, setShowBanner] = useState(false);
-  const res = pathAsResource(groupPath);
-  const [recipients, setRecipients] = useState([]);
-
-  useEffect(() => {
-    (async () => {
-    if (!res) { return; }
-    if (!group) { return; }
-    if (group.hidden) {
-      const members = _.compact(await Promise.all(
-        Array.from(group.members)
-          .map(s => {
-            const ship = `~${s}`;
-            if(s === window.ship) {
-              return Promise.resolve(null);
-            }
-            return props.api.contacts.fetchIsAllowed(
-              `~${window.ship}`,
-              'personal',
-              ship,
-              true
-            ).then(isAllowed => {
-              return isAllowed ? null : ship;
-            });
-          })
-      ));
-      if(members.length > 0) {
-        setShowBanner(true);
-        setRecipients(members);
-      } else {
-        setShowBanner(false);
-      }
-
-    } else {
-      const groupShared = await props.api.contacts.fetchIsAllowed(
-        `~${window.ship}`,
-        'personal',
-        res.entity,
-        true
-      );
-      setShowBanner(!groupShared);
-    }
-    })();
-
-  }, [groupPath]);
+  const {
+    api,
+    showBanner,
+    setShowBanner,
+    group,
+    groupPath,
+    recipients
+  } = props;
 
   const image = (props?.our?.avatar)
   ? (
@@ -88,7 +34,7 @@ export const ShareProfile = (props) => {
       <Sigil
         ship={window.ship}
         size={16}
-        color={`#${uxToHex(props?.our?.color)}` || "#000000"}
+        color={!!props.our ? `#${uxToHex(props.our.color)}` : "#000000"}
         icon />
     </Row>
   );
